Update reactive messaging repeats for MP7
