+++ dev/io.openliberty.microprofile.reactive.messaging.internal_fat/fat/src/io/openliberty/microprofile/reactive/messaging/fat/kafka/validation/KafkaValidationTests.java
@@ -34,7 +34,7 @@ import componenttest.custom.junit.runner.FATRunner;
 import componenttest.rules.repeater.RepeatTests;
 import componenttest.topology.impl.LibertyServer;
 import io.openliberty.microprofile.reactive.messaging.fat.suite.KafkaTests;
-import io.openliberty.microprofile.reactive.messaging.fat.suite.ReactiveMessagingActions;
+import com.ibm.ws.microprofile.reactive.messaging.fat.repeats.ReactiveMessagingActions;
 
 @RunWith(FATRunner.class)
 @AllowedFFDC
@@ -45,11 +45,9 @@ public class KafkaValidationTests {
     @Server(SERVER_NAME)
     public static LibertyServer server;
 
+    //Repeats for RM 3.0 tests only. MP 2.0 is not tested as it has RM 1.0
     @ClassRule
-    public static RepeatTests r = ReactiveMessagingActions.repeat(SERVER_NAME,
-                                                                  ReactiveMessagingActions.MP61_RM30,
-                                                                  ReactiveMessagingActions.MP50_RM30,
-                                                                  ReactiveMessagingActions.MP60_RM30);
+    public static final RepeatTests r = ReactiveMessagingActions.reactive30Repeats(SERVER_NAME);
 
     @BeforeClass
     public static void setup() throws Exception {
