+++ dev/io.openliberty.microprofile.reactive.messaging30.internal_fat_tck/bnd.bnd
@@ -25,11 +25,13 @@ javac.target: 11
 Require-Capability: osgi.ee; filter:="(&(osgi.ee=JavaSE)(version=11))"
 
 tested.features: \
+    cdi-4.1,\
     concurrent-3.0,\
     mpconfig-3.1,\
     mpmetrics-4.0,\
     mpmetrics-5.1,\
-    servlet-5.0
+    servlet-5.0,\
+    servlet-6.1
 
 # To define a global minimum java level for the FAT, use the following property.
 # If unspecified, the default value is ${javac.source}
@@ -39,4 +41,5 @@ tested.features: \
 #      or binaries from Artifactory (e.g. org.apache.derby:derbynet)
 # For all project names that match the pattern "*_fat*", dependencies for junit,
 # fattest.simplicity, and componenttest-1.0 will be automatically added to the buildpath
--buildpath:
+-buildpath: \
+    io.openliberty.microprofile.reactive.messaging.internal_fat.common.jakarta;version=latest
