+++ dev/io.openliberty.microprofile.reactive.messaging30.internal_fat_tck/build.gradle
@@ -1,3 +1,5 @@
 apply from: '../wlp-gradle/subprojects/maven-central-mirror.gradle'
-
+dependencies {
+    requiredLibs project(path: ':io.openliberty.microprofile.reactive.messaging.internal_fat.common', configuration: 'transformed')
+}
 addRequiredLibraries.dependsOn addJakartaTransformer
